Add files via upload
Working on creating a drop down menu to let user select a category and display the news of the category. No progress in this stage. now data is in view.py in the home folder. It cannot go to the <script> seesion in home.html. Tried to store the data in json file but still no clue on sending it back to the <script> session.