Add files via upload
views.py: added search function.
created home/form.py for url input
when users submit a url, the predict function will run to predict the category of the news and display the result in the home page.

code for import news information to the place in the home function in view.py. When user accesses home page, the code will run and load data from csv file to the database. However, the content of news may not be accepted by the database and will cause errors. (’) and (=) are symbols that the database is not accepted.

After modifying code in search function in views.py and prediction.py. The news can be imported to the database. The same problem as the demo part is that the date is not extracted. Also, categoryid is not correctly used